refactor: CandleStick component
