+++ src/components/CandleStick/options.ts
