+++ src/components/CandleStick/styles.module.scss
